VCL-1045 - Method of encrypting sensitive database entries
addomain.php: modified AJsaveResource: added check for getCryptKeyID returning NULL

siteconfig.php: removed TODO comment

utils.php:
-modified checkCryptkey: updated comments
-modified prettyLength: (unrelated to this JIRA) cast $minutes / 60 to int

vm.php: modified AJupdateVMprofileItem: added check for getCryptKeyID returning NULL

+++ web/.ht-inc/utils.php
@@ -646,10 +646,11 @@ function checkCryptkey() {
 			return;
 	}
 
-	# no id file or no matching entry in cryptkey, create new key
+	// if no id file and cannot write to cryptkey directory, return
 	if(! is_writable($filebase))
 		return;
 
+	# no id file or no matching entry in cryptkey, create new key
 	$keyfile = "$filebase/private.pem";
 
 	$_algorithm = constant("OPENSSL_KEYTYPE_" . ASYMALGO);
@@ -10023,7 +10024,7 @@ function prettyLength($minutes) {
 	elseif($minutes == 60)
 		return i("1 hour");
 	elseif($minutes % 60 == 0)
-		return $minutes / 60 . " " . i("hours");
+		return (int)($minutes / 60) . " " . i("hours");
 	else {
 		$hours = (int)($minutes / 60);
 		$min = (int)($minutes % 60);
